requirements.txt
flask==2.3.2
sqlalchemy==2.0.15
pandas==2.0.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
numpy==1.24.3
matplotlib==3.7.1
plotly==5.14.1
lxml==4.9.2
beautifulsoup4==4.12.2
gunicorn==20.1.0
gevent==22.10.2
